[#168303753] remove react-native-iap from Android build
react-native-iap is somehow triggering a problem in the library
react-native-netinfo that requires AndroidX.

https://github.com/react-native-community/react-native-netinfo/issues/107

Since we don't currently support IAP in Android, the easy solution is to remove
the react-native-iap code from the Android build.

Revert this commit when we update to React Native 0.60 (#168304197) and want to
allow IAP on Android through Google Play (#168313664).