[#175792142] patch react-native to fix missing iOS images
Add patch-package to project, enabling the patching of files in the
node_modules directory. As the node_modules directory is unversioned,
this provides a convenient way to apply versionable updates/bugfixes to
various node packages while waiting for an official fix - any patches
(in standard diff format) may be added to the patch/ directory, and will
be applied when building the app.

Add the following patch to react-native, fixing the missing image issue:
https://github.com/facebook/react-native/commit/123423c2a9258c9af25ca9bffe1f10c42a176bf3

(Patch may be removed upon updating to react-native 0.63.2 or greater.)