[#175800398] removed support for showing notifications in the app
Firebase 6 dropped support for showing notifications when the app is in the
foreground:

https://rnfirebase.io/migrating-to-v6#notifications

We only used this to show new issue notifications if the user was currently in
the app, which is unlikely and not particularly useful, so removing it for now.

If we want to show notifications in the app at some future time (e.g. to support
bulk messaging or user-to-user messaging), Firebase recommends these projects:

- https://notifee.app
- https://github.com/wix/react-native-notifications
- https://github.com/zo0r/react-native-push-notification