[#175800398] fix crash rendering subscription products
Upgrade react-native-iap to 5.2.7 to fix a crash caused by the fetched products
missing various attributes, including subscriptionPeriodUnitIOS, which leads
to an attempt to index an undefined variable.

https://rollbar.com/FrontPorchForum/Mobile/items/98/
https://rollbar.com/FrontPorchForum/Mobile/items/106/

https://github.com/dooboolab/react-native-iap/issues/1203

Note this crash only affects some devices or iOS versions.  Emily Potter reports
a crash on an iPhone 8 running 13.1.3, but success on an iPhone 7 running iOS 13.7.