[#177966351] warn iOS users if no app set to handle mailto links
If an iOS user has no application configured to handle mailto links, an
exception will be thrown when clicking a "Reply to Author" button.

This may also be triggered if a User selects Cancel when presented with
the "which app would you like to open [email links with]" option screen.

If a user who has no application available to open mailto links tries
clicking one, respond with an appropriate alert.

If a user clicks the Cancel button when asked to select an application,
take no further action (don't attempt to open the link).