[#177699335] allow null currentAreaId & hasAreaAccess values
In d743143, we switched currentAreaId from 0 to null in situations where
a member had access to no areas. (E.g. all their profiles have been
deactivated.)

Now that null is a permissable value for currentAreaId, the required
modifier needs to be removed from the associated property types.

Additionally, even though hasAreaAccess is initialized with a default
value of true in src/common/areas/slice.js, it appears to be null upon
initial render of the Forum tab. Allow that value to be null as well, to
avoid a warning.