[#177966320] improve axios error stack traces
The stack traces of errors resulting from failed asynchronous axios
requests do not include the actual code that called an axios request in
the first place. This makes debugging very difficult - use an axios
interceptor to append the original stack trace so that the code which
triggered the failing axios request may be determined via rollbar.