[#178646335] added back button to our webview component
We use our WebView component (which is built on top of react-native-webview)
to navigate through the directory.  During that navigation, it's useful for
users to be able to go back to the previous page.

Our webview now adds a back button to the app header whenever the location
stack has more than one entry.  Store the stack of visited urls as an array
in state to support this feature, since the react-native-webview's history
is broken by updating the webview source whenever a url changes.

NOTE:

This does not work in the dev env for /d/{profile id} pages, but does work for
other pages in the dev env and works for all pages in the staging env.  In the
dev env, the request.url == "about:blank" in the onShouldStartLoadWithRequest
callback, which short circuits the logic.

I'm not sure, but I think Turbolinks may be confusing the underlying webview.