[#179986817] downgrade from jetifier 2.0 to 1.6.8
Removed jetifier 2.0 dependency.  Instead, react-native will pull in jetifier
^1.6.2 (currently 1.6.8).

This removes a jetifier deprecation warning that AppCenter is interpreting as
a build failure that aborts the Android build:

  ##[error]Bash wrote one or more lines to the standard error stream.
  ##[error]Jetifier is deprecated. Future versions of react-native CLI may not
  run it by default.

https://github.com/microsoft/appcenter/issues/887#issuecomment-832394207
https://github.com/mikehardy/jetifier/releases/tag/v1.6.8
https://github.com/react-native-community/cli/issues/1409

We don't directly depend on jetifier, and it will be removed as a react-native
dependency at some point, so no need for it to be listed in our package.json.