[#180003989] Fix error in Keyboard by patching react-native-tab-view
Add patch-package to project, enabling the patching of files in the
node_modules directory. As the node_modules directory is unversioned,
this provides a convenient way to apply versionable updates/bugfixes to
various node packages while waiting for an official fix - any patches
(in standard diff format) may be added to the patch/ directory, and will
be applied when building the app.

Patch may be removed after the error is fixed in react-native-tab-view.
https://www.npmjs.com/package/react-native-tab-view