[#181117012] WebView: allow POST requests through explicit URL param
React Native WebViews only send headers on the initial request. However,
we set a user agent header to let the Rails app know a request is coming
from the mobile app. To ensure headers are included in subsequent
requests, we apply a kludge where we update the WebView component's
state upon detecting a URL change, triggering a fresh request with
headers.

This has an undesirable side-effect of converting any POST requests into
GET requests. To avoid this, allow the Rails app to explicitly declare a
POST request via a ?method=post param.

(It would obviously be preferable to examine the request object. But
since the request object doesn't expose it's internal HTTP method, we'll
use the ?method=post param instead.)