[#181576470] verify connection status before alerting user
We use React's NetInfo to determine when a user's internet connection is
weak, and alert the user accordingly. In theory, the alert should
dismiss once internet connection is restored or once the user clicks the
"Retry Connection" button.

However due to a bug in NetInfo, NetInfo often fails to detect when
internet connection is restored, and the alert becomes undismissable.

To prevent this, always double-check connection status by trying to
ping the web application. This way, even when NetInfo fails to detect
restored internet connection, the user will still be able to click the
"Retry Connection" button to continue using the mobile app.