[#180255292] potentially fixed mail app not opening
On iOS, the Mail app did not open when clicking a "Reply to Author" button
but did when clicking an email address in a `mailto` link.  Possibly the code
in `isAppInstalled` is not properly checking for the presence of the Mail app,
so if it's not found default to linking to a `mailto` link.

Also added `mailto` to `LSApplicationQueriesSchemes` info key to satisfy iOS
warning message.