[#180255292] only use side drawer menu for certain tabs
The side menu that shows a list of areas is most relevant to the Forum tab,
and shouldn't be shown at all on the directory and calendar tabs which need
to use the header left space for a back button.

Moved the side menu out of the root app container and into the authenticated
tabs component.  Removed the react context object that was used to track and
manage its state, and instead use basic react component state.  This state
can be manipulated by the different tabs to enable or disable the side menu.