[#185883871] fixed applying (and testing) max height in `SizedImage`
In the `SizedImage` component that resizes an image loaded from a URL using some
configured max width and/or height, the algorithm failed to apply a max height
when no max width was set (because `maxWidth` is undefined when the component
is first loaded).  Fixed this.

Also fixed the test, which wasn't asserting the width and height styles correctly,
and expanded the test cases.