[#186624203] handle forum URLs that include area slug instead of area id
The `appSettings` slice now includes a map of area slugs to ids used for
converting area slugs from forum URLs to area ids, which are expected in
the `Forum` view and elsewhere.

The `WebView` handles this conversion.  It would be preferable if the `Forum`
component could handle either an `areaSlug` or an `areaId` route param,
but that component has complicated (messy?) prop and route param handling
already.

Removed `parseDeepLink` since it's not in use.