[#186851356] fixed nav tab labels overlapping icons
On an iPad or other layout where the navigation tab label is on the same
horizontal row as the tab icon (as opposed to below the icon), the label
was overlapping the icon.

Fixed by removing the special `NavLabel` element from the label, and instead
pass just a string and set the label styling at the `Tab.Navigator` options.

- https://github.com/react-navigation/react-navigation/issues/7383
- https://reactnavigation.org/docs/upgrading-from-5.x/#the-tabbaroptions-prop-is-removed-in-favor-of-more-flexible-options-for-bottom-tabs

This bug has probably affected us for years.