[TECH-6504] updated to latest version of `rollbar-react-native` and `rollbar`
The old version wasn't compatible with Xcode 16 and caused a build error
similar to this one:
https://github.com/getsentry/sentry-react-native/issues/3883

The new version is technically in beta, but has been so for more than a year,
so is hopefully stable.

The new version requires iOS 14+. We previously supported iOS 13+, but it's
about time for us to bump that minimum version anyway.

The new version has the same API as the old version except for configuration
of the client. Migration doc:
https://docs.rollbar.com/docs/migrating-to-v100