[TECH-6504]-related: replaced metro global packages with yarn workspaces
We use short paths like `common/utils/rollbar` as an alternative to long relative
paths for js imports. Previously, this relied on metro global packages, but the
metro devs no longer recommend that option (because it's incompatible with other
source code tools).

Instead, we now use yarn workspaces, which create symlinks in `node_modules` to
our local src directories.

This necessitated renaming these packages from `@core` to `@fpf/core` because
yarn otherwise didn't like an @ symbol in the name. That change has the side
benefit of making it obvious that these are fpf packages.

- https://classic.yarnpkg.com/lang/en/docs/workspaces/
- https://github.com/facebook/metro/issues/1166#issuecomment-1878955929