[TECH-6504] removed `SafeAreaView` from main app layout
This view is supposed to ensure that any phone screen cutouts or OS UI elements
do not obscure the app's own content. However, when applied to the main app
layout, it was creating too much whitespace at the top of the iOS view.
Removing it doesn't seem to have any negative consequences.

For the `FullScreenWizard` view, removing it caused the top of the wizard view
to be obscured by the iPhone notch, so instead I updated it to use a more
recent API (`useSafeAreaInsets`).