[TECH-6504] alias for `yarn list` to fix AppCenter error
AppCenter is trying to call `yarn list` (at least for the Android build), and
that command no longer exists in yarn 2+. Work around by implementing that
command as a script that does nothing. (It might be nice for it to call
`yarn info`, which displays similar info, but AppCenter is passing the arg
`--depth=0` to it, which `yarn info` doesn't support).

https://github.com/microsoft/appcenter/issues/2134#issuecomment-954067024