[TECH-6504] fixed broken elision in TOC lines on Android
On Android, the TOC is not eliding post titles correctly. Instead, it's showing
a single line without elision, and (with some very careful finger work) you can
scroll each title.

This appears to be a react native bug, and the only way I can fix it is to make
these text titles not user selectable.

https://linear.app/front-porch-forum/issue/TECH-6504#comment-6a05c442