[TECH-6950] don't set navbar title to url
We sometimes use the document title from the page loaded in the webview to set
the navbar title, e.g. the forum tab's title becomes "Montpelier Forum" when
`/montpelier/forum` has been loaded.

Android briefly sets the document title to the document path, and we don't want
to show this url in the nav bar.