[TECH-6964] potential fix for blank tabs
The WebView is run as a separate process (in both iOS and Android, I think)
that can halt without the full FPF mobile app also halting. When that happens,
the webview tabs (e.g. forum) won't load and appear blank white.

Potentially fix this bug by using webview callback options that can reload
the webview after a crash.

However, this bug is hard to replicate, so I'm not certain if this bug is
actually fixed.

There's a long-standing react native webview issue, but it was closed without
being fixed:

https://github.com/react-native-webview/react-native-webview/issues/3062