[TECH-6978] run jest tests under Android platform as well as iOS
Previously, our unit tests only ran for the iOS platform only (I think that's a
react native or jest default, but not sure). As a result, any Android specific
code was skipped by tests.

This commit adds jest config files for ios and android and runs the test suite
with both configs successively. It also fixes a few tests for android.