[TECH-6948] tap on tab button when tab is focused reloads webview
(except for Compose, which scrolls to top so that a user doesn't accidentally
lose their in progress post).

NOTE: the change to `return false` from `navigateForRequest` breaks the forum
loading for some reason (possibly related to redirects), so this commit is
unresolved.

---

This tab function will allow users to return to the root url of a tab more easily.

Our `WebView` component now takes slightly different props:
- `rootUrl` for the home page of the tab
- `initialUrl` for the currently loaded page of the tab
- `headers` for optional request headers (necessary for auth)
- `reloadRootOnTabPress` and `scrollTopOnTabPress` to control behavior

The url and headers props replace the  single `source` object prop.

Also reorganized the code in WebView a little for (maybe) readability.