[TECH-6948] simplified WebView stack tracking
Previously, our wrapper around the WebView tracked its stack and current url in
state in order to ensure that custom headers were passed along with each new
request. That architecture was based on this guide, and I think the need to
include basic auth headers with each request (for staging):

https://github.com/react-native-webview/react-native-webview/blob/master/docs/Guide.md#working-with-custom-headers-sessions-and-cookies

However, that was quite complicated, and it seems unnecessary now that we don't
require basic auth for mobile app requests. The custom `authorization` header
is also necessary, but I guess our Rails app is happy enough for it to only be
present in the initial request from the WebView...

So this new code leaves stack management up to `react-native-webview`, and only
changes the source url when necessary to:
- reload the current page
- reload the "root" page
- navigate to a different page requested by a different tab