[TECH-6948] fixed opening non-whitelisted fpf urls from webview
When the webview's `onShouldStartLoadWithRequest` returns false for a
non-whitelisted frontporchforum.com url in order to cancel a request, we
need to use `Linking` to open the url in a browser.

`navigateWithToken` will do so with the added benefit of attaching a
temporary auth token so that the user is logged in in the browser,
and `openFpfUrl` helps by adding query params to the url to identify
the traffic in plausible as originating from the mobile app.