[TECH-7034] update Android dependencies
The Android Gradle plugin still needs to be updated, according to the
following warning received when running `yarn android` to build the app:

---

WARNING: We recommend using a newer Android Gradle plugin to use compileSdk = 36

This Android Gradle plugin (8.6.0) was tested up to compileSdk = 35.

You are strongly encouraged to update your project to use a newer
Android Gradle plugin that has been tested with compileSdk = 36.

If you are already using the latest version of the Android Gradle plugin,
you may need to wait until a newer version with support for compileSdk = 36 is available.

For more information refer to the compatibility table:
https://d.android.com/r/tools/api-level-support